Code

+++ src/main/java/com/driver/services/BlogService.java
@@ -26,7 +26,7 @@ public class BlogService {
 
     public List<Blog> showBlogs(){
         //find all blogs
-        List<Blog>  blogList = blogRepository1.findAll();
+        List<Blog> blogList=blogRepository1.findAll();
         return blogList;
 
     }
@@ -34,29 +34,50 @@ public class BlogService {
     public void createAndReturnBlog(Integer userId, String title, String content) {
         //create a blog at the current time
 
-        Blog blog = new Blog();
         //updating the blog details
 
-        User user = userRepository1.findById(userId).get();
         //Updating the userInformation and changing its blogs
 
+        User user=userRepository1.findById(userId).get();
+
+        Blog blog=new Blog(title,content);
+
+        blog.setUser(user);
+
+        List<Blog> blogList=user.getBlogList();
+        blogList.add(blog);
+        user.setBlogList(blogList);
+
+        userRepository1.save(user);
     }
 
     public Blog findBlogById(int blogId){
         //find a blog
-        Blog blog = blogRepository1.findById(blogId).get();
-        return blog;
+        return blogRepository1.findById(blogId).get();
 
     }
 
     public void addImage(Integer blogId, String description, String dimensions){
         //add an image to the blog after creating it
-        Blog blog = blogRepository1.findById(blogId).get();
-        imageService1.createAndReturn(blog,description,dimensions);
+        Blog blog=blogRepository1.findById(blogId).get();
+
+        Image image=imageService1.createAndReturn(blog,description,dimensions);
+
+        image.setBlog(blog);
+
+        List<Image> imageList=blog.getImageList();
+        if(imageList==null) imageList=new ArrayList<>();
+        imageList.add(image);
+        blog.setImageList(imageList);
+
+        blogRepository1.save(blog);
+
     }
 
     public void deleteBlog(int blogId){
         //delete blog and corresponding images
-        blogRepository1.deleteById(blogId);
+        if(blogRepository1.findById(blogId).isPresent()){
+            blogRepository1.deleteById(blogId);
+        }
     }
 }
